requirements.txt
# Core Dependencies
flask==2.3.3
transformers>=4.36.0
sentence-transformers>=2.3.0
torch>=2.1.0
faiss-cpu==1.7.4
numpy>=1.24.3
scikit-learn>=1.3.0
requests>=2.31.0

# Text Processing
nltk>=3.8.1

# Utilities
python-dotenv>=1.0.0

# Explicitly specify compatible huggingface_hub version
huggingface_hub>=0.19.0,<0.25.0